dash_sim/requirements.txt
dash>=2.14.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
Pillow>=10.0.0
svgpathtools>=1.6.1
scipy>=1.11.0
shapely>=2.0.0
geopandas>=0.14.0
pyyaml>=6.0
